content_review_tool/requirements.txt
google-generativeai
python-dotenv
pydantic
pillow
